cleanup(generator): remove obsolete parameter comment substitutions (#11021)
A couple of the ad hoc substitutions we apply to the comments for
method_signature parameters are no longer needed (the ones from
google/pubsub/v1/schema.proto), so remove them.

How do we know they're obsolete?  By making it a fatal error for a
substitution to fail to match.  This will make it easy for us to
remove dead cruft.