doc: do not mention Conan (#11036)
After trying to get Conan working for 3 months, I think we should stop
mentioning it as a viable package manager.  I will continue the work to
get it working, but it seems too brittle and hard to use.  Mentioning
seems to be interpreted as an implicit recommendation, and I cannot say
that would recommend it.