feat(GCS+gRPC): more efficient CRC32C computations (#11070)
Move the CRC32C and MD5 computations to the `GrpcClient` and
`RestClient`. In the case of gRPC this allows us to compute the CRC32C
checksum for each `Write()` message and then concatenate this checksum
with the CRC32C checksum computed for the full object.  That saves one
of the most expensive computations in the gRPC uploads.