doc(common): remove undefined TODO from StatusOr
There was an obsolete and undefined `TODO()` in the documentation for
`StatusOr`. First, TODOs like these do not belong in the public
documentation. Second, the TODO was obsolete since `StatusOr` uses
`absl::optional<>` and therefore works with whatever `absl::optional<>`
supports.

I introduce a `static_assert()` to verify `T` is not a reference type.
That was never supported, and this should create a more readable error.

Finally, I made some editorial changes to the comments.