feat(spanner): add the final pieces for the RouteToLeaderOption
The final glue to add a routing header to Spanner RPCs that should be
served in the leader region.

For the time being, the boolean option defaults to off (set as false),
but that default can be changed with a suitably-positive value for
`${GOOGLE_CLOUD_CPP_SPANNER_ROUTE_TO_LEADER}`.  The conditional code
can be removed to restore the defaults-to-on behavior.  See #11111.