impl(bigquery): unify bigquery request/response logging format (#11172)
It is a feature if request/response logging uses a common format no matter
whether the message types are protoc generated or not.  So, update the
`DebugString()` functions for `GetJobRequest` and `ListJobsRequest` to
produce that common format, as is they had been defined as messages like ...
```
message GetJobRequest {
  string project_id = 1;
  string job_id = 2;
  string location = 3;
}

message Projection {
  string value = 1;
}

message StateFilter {
  string value = 1;
}

message ListJobsRequest {
  string project_id = 1;
  bool all_users = 2;
  int32 max_results = 3;
  google.protobuf.Timestamp min_creation_time = 4;
  google.protobuf.Timestamp max_creation_time = 5;
  string page_token = 6;
  Projection projection = 7;
  StateFilter state_filter = 8;
  string parent_job_id = 9;
}
```

In particular, `ListJobsRequest::min_creation_time` and `max_creation_time`
are now logged.

Also expand the tests to cover the `TracingOptions`, move some file locals
to the anonymous namespace, and remove some unused variables.