impl(GCS+gRPC): support OSS version of ctype=CORD
The OSS version of ctype=CORD will not support the `mutable_*()`
accessor (in our case that is `mutable_content()`). One needs to use
`set_content()` and `content()` exclusively. This PR adds more template
dispatch magic to use `*mutable_content()` when the protos do not have
`ctype=CORD` annotations (or the annotation has no effect), and use
`set_content()` when the protos have the annotation.