impl(GCS+gRPC): support OSS version of ctype=CORD
The OSS version of ctype=CORD will not support the `mutable_*()`
accessor (in our case that is `mutable_content()`). One needs to use
`set_content()` and `content()` exclusively. The workarounds to access
the private member functions fail to compile (even if unused), so we
need a compile-time flag and some heavy documentation.