feat(common): a "lastN in thread" logger
This is a variant of the "lastN" logger. It only dumps the lastN log
lines *in the thread* when the library or application creates a log line
over a minimum severity threshold. This is an improvement over the
"lastN" logger, which dumps the last N logs from all the threads in
similar circumstances.