ci(spanner): re-enable DataBoost samples but with workaround
The problem with the DataBoost samples was that there is no capacity
in the region used by our Spanner instances labelled as "samples:yes".

So, lets re-enable them, but with a specific allowance for seeing a
`kPermissionDenied` status code.  Then, they will automatically start
working when capacity is allocated, and after which we can remove the
workaround.

Use `//!` to keep the workaround out of the published snippets.

Related to #11201.