ci: publish-docs always run DocFX generation (#11315)
Running the `doxygen2docfx` tool in each PR will ensure we do not
regress w.r.t. supporting all the libraries in the tool. We also need to
run the tool to (eventually) upload the documents to start the
publication pipeline, so this is a good place to run them.

I had to change how the tool is integrated into CMake. It cannot be a
"feature" (an element in the `GOOGLE_CLOUD_CPP_ENABLE` list) because
CMake needs to configure the `docfx/` directory before the
`google/cloud/` directory and the features always go afted
`google/cloud/`.