impl(docfx): iterate on TOC (#11319)
Start the table of contents (TOC) with the `indexpage` and use
`index.md` as the output file for that page. Doxygen uses `indexpage`
for the landing page, the cloud site uses `index.md`.

Move the namespaces up in the TOC, and use the C++ name (instead of the
Doxygen uid) as the name of the TOC entry.

Move the groups down in the TOC, and use the group title as the name of
the TOC entry.

Keep the TOC entries sorted.