fix(spanner): propagate Options through SessionPool callbacks (#11344)
`BatchCreateSessions()` calls made during the `SessionPool` callback
chain depend on the prevailing options to determine whether the RPC
should be routed to a region leader.  So, we need to ensure that the
construction-time options are propagated through those callbacks by
instantiating `OptionsSpan` objects when appropriate.

Add similar `OptionsSpan` objects to `session_pool_integration_test.cc`,
which uses "FriendForTest" access to reach into `SessionPool`, bypassing
the normal span instantiation.

Add unit tests to verify that the "x-goog-spanner-route-to-leader"
header is now set appropriately when the `RouteToLeaderOption` is both
enabled and disabled.

Add `RouteToLeaderOption` to the `SessionPoolOptionList`, marking it
as a valid option to pass to `spanner::MakeConnection()`.

Also add a slight optimization to `SessionPool::Grow()`, which skips
even trying to create sessions when `ComputeCreateCounts()` returns an
empty vector.