impl(docfx): generate full table of contents (#11493)
Change the table of contents generation to include namespaces, classes,
structs, functions, enums, typedefs, etc. Group elements by "type", so
structs appear together. The grouping matches the groups in the
right-side navigation for a namespace or class. Compound elements
(namespaces, classes, structs) have an 'Overview' link to point to the
element description. There is no way to have a grouping that is also a
link, that is, a group for "all the stuff in class Foo that also links
to Foo" is not a thing.

Links to enums and typedefs are not pointing anywhere at the moment,
that will be fixed in separate PRs.