impl(docfx): iterate on table-of-contents
For functions and enums (and in general for `<memberdef>` elements) we
do not want to use the fully-qualified name, it is just too cumbersome.
Likewise, items in the table of contents without subitems (like
functions, enum values, and typedefs) we don't need to create an
`Overview` section, just use the item as its own link.