impl(docfx): generate separate file for enums (#11497)
The tooling at google expects enums to be generated as separate files.
This adds a small loop in `main()` to generate the files. We need to
remove the enum definitions from the previous yml files, or we get
duplicate definitions. And to make the enumvalues visible we need to
list them as children of the enum.