impl: prepare for next Protobuf release (#11518)
On Windows, `GetCurrentTime()` is a macro. The current and previous
releases of Protobuf undefine this macro in one of its headers.  That
was probably a bug in Protobuf.  We use `TimeUtil::GetCurrentTime()`.
With the next release of Protobuf that will expand the macro. We either
need to prevent the macro expansion, using something like
`(TimeUtil::GetCurrentTime)()`, or we need to avoid the function
altogether.