cleanup(ci): log cmake/ctest command lines
In #11515 I removed `io::run` from the cmake command in the "m32"
build, in the interests of consistency with the other cmake builds,
which don't use logging.

Here, we still maintain consistency but by adding `io::run` to all
the other cmake/ctest commands instead (well, at least those in the
top-level build scripts).  This makes the build output easier to
follow.