ci: disable `dlp` if Protobuf >= 23 and GCC == 7.5.0
Protobuf >= 23.0 generates code that triggers an internal compiler error
with GCC == 7.5.0. Earlier versions of GCC appear Okay, and later
versions also appear Okay. The only distro shipping with GCC == 7.5.0 is
Ubuntu 18.04. That is not long for this world, and there is not much we
can do to fix Protobuf or GCC.