feat(generator): generate repository metadata files (#11555)
The documentation index is based on the `.repo-metadata.json` files
committed to our GitHub repository. With this change, the generator
will automatically create and update these files.

We need to get some information from the service config YAML file, so
this change introduces changes to find and parse the file. There are
other cool bits of information on that file, which we can use later.

The generator only creates a `.repo-metadata.json` file if all the
necessary information is found. This can be useful to manually create
the files when the information is missing, as the generator will not
overwrite them.