impl(common): support extracting a typed option from an Options
```
absl::optional<typename T::Type> internal::ExtractOption<T>(Options&);
```
Extracts and returns the value for `T` from the options, or nullopt if `T`
was not set.

This is intended for code paths that want to consume an option, for example
in a client call that passes the option's value via conventional mechanisms,
rather than though an `OptionsSpan`.