impl(docfx): add group childrens to description (#11619)
I could not figure out how to represent the children of a group as
`children` elements in DocFX. It appears that having them as children of
the namespace and the group does not match the expected data model. With
this change they are listed, and linked, which is enough for our
purposes.