feat(generator): improve method comments
Add more details about the returned types for different method
"category" (paginated, long-running, bidir, etc.). Include links to the
proto message and to the mapping rules. Explain the role of `StatusOr`
and `future` in these RPCs.

Also include a longer description of the `request` parameter, including
links to the Protobuf mapping rules and the original proto message.

I disabled formatting around the comments for each method. It was too
hard to create stable reference-style links, and I think we don't want
to mess with the newlines introduced in the proto comments.