fix(rest): support rewinds in libcurl
We configure libcurl to get the data for PUT and POST requests using
callbacks. Sometimes libcurl may have sent part of the data and needs to
resend it. The documentation mentions "multi-pass authentication
methods", as well as reusing connections where libcurl detects they were
closed after some data is sent. There are cases on the Internet about
chasing HTTP redirects (3xx responses) too. In any case, we have seen
these as flakes in the tests.

I took the belt and suspenders approach. Try to handle the rewind
requests, but if unhandled treat the error as retryable, because they
are.