refactor(bigtable): future-proof ReadRows mocking (#11725)
Add a new struct `bigtable::ReadRowsParams` to consolidate configuration parameters for `Table::ReadRows()`. This future-proofs the growing of the API. It allows the library to stay up to date with the latest and greatest capabilities from Bigtable.

Adds a `DataConnection::ReadRowsFull(ReadRowsParams params)` for mocking. All calls to `Table::ReadRows(...)` flow through this method.

Customers that mock the library should use `MockDataConnection::ReadRowsFull(...)` in their tests.