test(storage): deflake samples vs. production
Many of the samples create buckets. This is a rate limited operation. To
make these operations reliable against production we need longer
timeouts for the tests, as they will need to retry for longer. We also
should start the backoff with a larger interval, as 1 second is not
enough for the rate limit to reset. We do something similar in the
integration tests with far fewer flakes.