impl(bigquery): InsertJob api prep work - Part 1 (#11753)
InsertJob api requires additional fields to be returned in the Job response.

This PR and subsequent PRs will add the missing fields and info to Custom Job classes, as needed by the InsertJob api.
These were not added in the initial PRs for the Job resource, as they were not needed for the read-only apis.

I have divided the above changes into 4 PRs. The first two will add the missing fields and sub classes. The remaining two will mainly add the DebugString() member methods and unit tests, for the logging decorator .

This PR includes the following changes:

1) Adds new JobConfigurationQuery class with following sub classes: SystemVariables, StandardSqlField, StandardSqlDataType, StandardSqlStructType, Struct and Value
Fields are recursive and self-refrential. Protobuf oneof functionality is simulated by absl::variant.

2) Modifies JobConfiguration class with the query class abopve

3) Json Parsers and Unit Tests for all.