feat(generator): trim overly long field comments (#11859)
The documentation in some fields requires special hacks to get Doxygen
to parse them more or less correctly. For the next service I am planning
to generate (`aiplatform`) this was getting out of hand. I think just
preserving the first paragraph and linking the protobuf message
documentation is enough, so that is what I did.