ci: increase timeout for macOS builds (#11965)
We know that a build with a cold cache will take more than 4h (240m). We
also know we have a few of these builds every month. Increase the
timeout to 8h. This should be enough as we never need a second build for
these sporadic timeouts.

This is not without tradeoffs. We also get broken builds from time to
time that simply stall without any progress. To minimize the impact of
these, this PR only increases the timeout for the builds that have timed
out while still making progress.