ci: use remote cache for CMake on GCB
Use `sccache` for CMake-based builds running on Google Cloud Build
(GCB). These builds will store the cache in Google Cloud Storage, so
partial builds will warm up the cache. My experiments show that the
build is approximately as fast with `sccache` as with `ccache` (plus
upload and download time to preserve and restore the cache). In some
cases `sccache` was faster. And it is simpler to manage.