ci: cleanup CMake flags for quickstart builds (#11990)
I missed this build in the cleanup to take advantage of CMake >= 3.13.
We were also passing `-DGOOGLE_CLOUD_CPP*` flags to the quickstart
builds. These have no effect because we use
`google/cloud/*/quickstart/CMakeList.txt` as the top-level CMake file in
these builds, and these flags are only defined in the ./CMakeList.txt
file.