feat: support split packages (#12049)
With this change package maintainers can split `google-cloud-cpp` into
smaller packages. Some package ecosystems have constraints on the build
time for the packages, or on the size of the packages.
`google-cloud-cpp` is exceeding some of these limits (notably build
times on Conda). The solution is to create a smaller package with the
core libraries, and then one or more packages with a subset of the
features built on top of this core.

In addition to the necessary support in CMake, this introduces a build
to test this configuration.