ci: more robust vcpkg cache for GHA build (#12077)
The vcpkg binary cache uses the compiler executable SHA as part of its
hash. We need to include these files in the GHA hash or the GHA cache is
treated as "valid" even though it is no longer usable.

Save the vcpkg cache as soon as it builds correctly. This preserves the
vcpkg cache when our build fails but building the deps succeeded.