ci(gha): restructure external accounts build
Restructure the external-accounts build to support additional builds in
the near future. All the steps to prepare the VM running the build are
in the workflow file. The steps to compile the code are in a shell
script, which follows the structure of analogous scripts for the
GCB-based builds.

This build is still "special" in that it runs one integration test from
the workflow file. We may want to refactor that in the future.