ci(gha): create configuration for Windows builds
As usual, we want to test the builds with Bazel and CMake. We also
support MSVC 2019 and MSVC 2022, so we want to test with both compilers.
And with MSVC there is a big difference between the C++ library in Debug
vs. Release mode, so we need to test both. There is no need to test all
these things with Bazel and CMake. So Bazel gets one build and CMake
gets 4.  We also need to shard the builds.