ci(gha): macOS PRs only compile Core shard
I would have wanted to write `if: ${{ matrix.shard == 'Core' || !PR }}'
to only run `Core` matrix on pull requests, but `matrix` is not
available for `job.if`. So instead we receive the list of shards from
the top-level job, based on the name of the event. But then we need a
way to compute the `features` for each shard. The mini-language for
matrices is not powerful enough to express a dynamic matrix like this,
so we compute the features in a helper step.