ci(gha): run smaller matrix on PRs
After this change is merged, PRs will only run a small matrix on GHA.
Basically the hand-crafted will be tested with CMake on macOS and
Windows.

A merge to the main branch runs the full build.

If we deem it necessary, we cah use the `gha:full-build` label to
trigger a full build of the branch.