refactor(storage): split `RetryClient` tests (#12321)
The tests for `storage::internal::RetryClient` were mixed with the tests
for `storage::Client`. This splits the tests to their own files. In a
future PR these tests will get a bit more interesting.

I also simplified the implementation of these tests. They now use more
canonical idioms to mock and expect results, and they all assume the
operations are idempotent. The existing tests for the retry loop and
idempotency already cover the case when an operation is not idempotent.