impl(GCS+gRPC): `GrpcClient` is a `GenericStub` (#12343)
And as a `GenericStub`, it receives the `rest_internal::RestContext` and
can propagate the idempotency token header to the service. In the case
of operations emulated via OCC loops (most of the *AccessControl
modifiers), we want the latest value after each iteration in the OCC
loop. That is, the idempotency token should not be propagated in this
case.