ci(spanner): stop using ephemeral ports for emulator with cmake
Like the bazel builds, use a non-ephemeral port for the Spanner emulator
when building with cmake.  Originally, in the bazel case, we used a fixed
port so as to not invalidate the test cache.  But this meant that we also
used a fixed, non-ephemeral port when adding the second emulator invocation
(with HTTP support). Now we do the same with cmake for the second reason,
which (largely) eliminates the chance of "Address already in use" errors.

Fixes #11073.

Also make the two invocations look similar, correct some environment
variables names, and add the "--copy_emulator_stdout" option in the HTTP
case (in case that is useful in future debugging).