impl(common): add internal::RemoveOptions<T...>(Options)
Add a helper function to remove all options in an `OptionList`.  This
is intended as a companion to `internal::ExtractOption<T>(Options)`,
for use in services where the `Options` passed to a client operation
are split into two subsets: (1) those options passed explicitly to the
connection layer and beyond, and (2) the remainder that are passed
implicitly via thread-local state.

For example, in a `Client::Foo(..., Options opts)`, ...
```
  auto foo1_option = internal::ExtractOption<Foo1Option>(opts);
  auto foo2_option = internal::ExtractOption<Foo2Option>(opts);
  internal::OptionsSpan span(
      internal::RemoveOptions<ExplicitOptions>(std::move(opts)));
  return conn_->Foo(..., {foo1_option, foo2_option});
```
where `ExplicitOptions` contains all the option types in category (1).