test(pubsublite): reduce quota requirements
Pub/Sub Lite topics consume at least 4MiB/s of throughput capacity per
partition. The default quota is 64 MiB/s. With 3 partitions a few
simultaneous runs of the test can consume all the quota.

This change reduces the number of partitions to 1, so we can run at
least 16 tests simultaneously.

The test was also mislabeled, and was running in most builds even though
it does not use an emulator.

We could also use Lite Reservations to further reduce the capacity
requirements per run, but that seems like a problem for another day.