feat(pubsublite): caller provides admin connection (#12422)
The Pub/Sub Lite publisher needs an admin connection to query attributes
of the topic at run-time. Initializing an admin connection is not
trivial, and the code was doing it subtly wrong. It is better to receive
the `*Connection` from the caller, so they can provide the right
`Options` in its initialization.