impl(GCS+gRPC): a helper for AsyncInsertObject (#12452)
A new class that performs a single asynchronous attempt to upload an
object given a buffer. The buffer is represented by an `absl::Cord`, but
this is not (necessarily) what the application will provide.

This class will be wrapped by `AsyncRetryLoop()` to handle transient
errors. It cannot be implemented as a simple function because we need to
emulate coroutines.