impl(GCS+gRPC): a class to represent write payloads (#12458)
We need a public API for object payloads. Applications may need to
examine the contents of the upload in their mocks, to ensure their code
is sending the right data.

We want the implementation of this type to be an `absl::Cord`, so we can
pass the data to gRPC and Protobuf without copying the contents.