chore(deps): tweak definition of libpfm dependency
Followup to #12568.

Remove the `str()` wrapper around `Label()`, which confuses the deps
cacher script `bazel/deps-cache.py`.

Also remove the `/download` suffix from the URL, that similarly confused
the deps cacher, and which appears to be superfluous.  This also removes
the need for the `type` specifier.

Finally, reorder the `maybe()` arguments to match the other invocations.
Provide an alternative URL to