impl: only decorators inject x-goog-api-client
For automatically generated REST-based clients we were injecting the
`x-goog-api-client` header twice: once in the generated
`*RestMetadataDecorator` class and once more in the
`rest_internal::CurlRestClient` class.

With this change the hand-crafted libraries inject the header
explicitly, and use the new `HandCraftedLibClientHeader()` to do so.
Meanwhile, the generated libraries inject the header only in the
decorator.