impl: make external/googleapis/renovate.sh a little smarter
Move some of the steps used to update the googleapis/googleapis
version from the recipe documented in the how-to guide, to the
`renovate.sh` script.

In particular, the script now updates the protodeps/protolists
files, runs the generate-libraries build, and creates git commits
ready to be pushed.

This is another step towards possibly automating the whole process.