impl(common): introduce an absraction for a Cloud location
Introduce `google::cloud::Location` to model a location resource
like `"projects/project-id/locations/location-id"`, and use this
in places where we previously worked directly with strings.

After this change has made it into a release, we can update the
quickstart programs (and the scaffolding code that generates them)
to use `Location`.

Part of #12631.