impl(GCS+gRPC): wrap streaming read RPC for storage (#12681)
Introduce a new class that plays the `*Connection` role for the
`AsyncReader` class. Unlike the raw streaming read RPCs, this returns
the types we want to expose to the applications (e.g.
`storage::ObjectMetadata` and `ReadPayload`) as opposed to the raw
protos. It will be used in tests mocking the `AsyncReader`, so it must
be public.