fix(GCS+gRPC): race condition in async `Insert()`
The code assumed the `OnWrite()` callback (and the follow-on `Write()`
call) would be invoked after it had a chance to prepare for the
`Write()` call.  Clever in that overlapped some I/O with CPU operations,
but wrong in that scheduling does not work like this.

I also changed the test to make the errors easier to interpret.