cleanup: remove `ci/etc/full_feature_list`
We had two sources of truth for the "list of all the libraries". This is
clearly undesirable, and it was time to fix it. I had to change some of
the CMake scripts, and took an opportunity to (in my opinion) simplify
them.

This makes `oauth2` a regular library, but then its `main.dox` must
follow the pattern for hand-crafted libraries so it is skipped by the
script to automatically update it.