ci: build script for developer workflows
This script is expected to support /most/ developer workflows. We will
be able to create a configuration file in $HOME or in the workspace to
build only a subset of the code, with clang-tidy, and to run (or not
run) the unit and integration tests associated with that code.

I expect this script will grow and be more customizable over time, as
the team adds features to match their workflow.