fix(storage): check ReadObject options at runtime
Some options for `ReadObject()` can only be freely combined. The
application intent is unclear when somebody says `ReadRange(1, 10)` and
`ReadLast(7)`. And it is impossible to represent this as two
instructions that the service can disambiguate.

We were checking for the presence of these options in the parameter
pack, but that does not work, because the options can be "set" with no
value (maybe at runtime choosing to set either `ReadLast()` or
`ReadRange()`). Adding insult to injury, I forgot to `std::decay_t` the
types in the parameter pack, so the compile-time check did not always
work either.

In hindsight we should have represented them by a `absl::variant<>` or some
other type that did not allow for illegal combinations, too late to do
that.