test(spanner): extend bad-session tests for streaming-read RPCs
Extend the "Session not found"-returning tests for `Read()` and
`BatchWrite()` to make an additional call to verify that the bad
session was not returned to the pool.  Streaming reads follow a
different model given that the error response is part of the
stream, so they deserve some additional coverage.