ci: enable `storage_grpc` with oldest deps (#13019)
Some missing headers in `storage/internal/grpc/stub.cc`.
Older versions of `absl::Cord::chunks()` cannot be used with
googletests's `ElementsAre()`.

The test for `storage_internal::CreateDecoratedStubs()`
required some tweaks: (1) to work with older versions of
gRPC they needed to have an active completion queue, and
(2) to work with newer versions they needed to use an
unreachable endpoint for the channels. Otherwise the
channel refresh timed out what is supposed to be a unit test.