test(GCS+gRPC): deflake stub factory test (#13104)
Change the factory helpers to avoid race conditions in the test. The
`GrpcChannelRefresh` object should be deleted after the stubs and the
corresponding channels. The test was destroying them in the opposite
order, this results in race conditions with older versions of gRPC.

I solved the problem by *not* creating a `GrpcChannelRefresh` that then
goes unused.

I think this was a test-only bug: the library does destroy these things
in the right order.