cleanup(bigtable): split out howto_mock_data_api bazel rule (#13189)
Using the same rule for `howto_mock_data_api` as the other
`bigtable_examples_unit_tests` means that the union of the
dependencies for all test are used by each.  This is causing
a problem for an important google_cloud_cpp consumer.

So, split out the `howto_mock_data_api` rule so that we can
minimize its dependency count.  This avoids the downstream
issue.