docs(spanner): reorganize Client::ExecuteQuery() docs
The overloads for `Client::ExecuteQuery()` used `@copydoc` to share
documentation, but this meant they had overly general specifications,
and included examples that did not apply to the overload at hand.

Let's instead give each overload an independent documentation block,
so that the text and examples are closely related to the behavior.