fix(common): returnable argument to `.then()`
The main motivation for these changes is to support:

```cc
future<T> F(future<T> f) {
  return f.then([](auto g) { /* stuff */; return g; });
}
```

This is common enough, and was broken for fairly obscure reasons.  I
completely changed the implementation of `.then()`. It was too hacky to
do something else, and I think the new implementation is simpler too.