cleanup(compute): shorten target names
Shorting the target names from `google_cloud_cpp_compute_*` to
`compute_*` reduces the path length for intermediate artifacts, such as
`.o` and `.obj` files. This prevents problems with MSVC, which is
limited to 255 characters in file names.

Note that the target name is used as the default name for the output
artifact (libraries in this case). This change overrides these defaults
to preserve backwards compatibility.