impl(generator): method variables for `request_id` (#13612)
This sets the method variables needed to support `request_id`-like
fields. As these variables need the service config YAML file, we need to
make the location of these file available to the code generator. We also
need to parse the file and propagate the results to the
`CreateMethodVars()` function.

We use the API index file in the googleapis repository to find the
location of this service config YAML file. That does not work for the
golden protos used in integration tests, so we need a way to override
it.