test(spanner): ignore contraint violations from CreateInstanceConfig()
If an CreateInstanceConfig() call fails with a constraint violation
it is undoubtedly the case that we chose a base config whose optional
replicas cannot all be added to the new, `USER_MANAGED` instance config.

It isn't worth trying to encode/discover constraint knowledge in our
tests or samples though, so we simply skip dependent tests/samples
upon a violation, and move on.

Resolves #11346.  Tested by forcing the selection of base config as
projects/cloud-cpp-testing-resources/instanceConfigs/regional-us-central1,
where "europe-west9" and "asia-south1" are known to be problematic
replica locations.