feat(storage): support HNS-enabled buckets
Add accessors and modifiers for the `hierarchical_namespace()` field in
a bucket. Parse the field in bucket metadata responses for JSON and
gRPC.  Add support to set this field on create, update, and patch
requests. Also for JSON and gRPC. Include and example (which doubles as
an integration test), and the usual unit tests.