impl(GCS+gRPC): helper functions to resume downloads
To resume a download we will need to create new `AsyncReaderConnection`
instances using the original request, but with (maybe) the generation
and the read offset updated to reflect any changes. This PR introduces a
few helper functions to update the request and a type to represent a
"factory" of `AsyncReaderConnection` instances.