impl(GCS+gRPC): a `ReaderConnection` with resumes (#13778)
We use a retry loop to *start* downloads. Sometimes a download may be
interrupted in the middle. When that happens we need to resume the
download from the last received byte. A simple retry loop won't help us
in this case. This PR introduces a class to resume downloads that are
interrupted.