test(spanner): enable more samples on the emulator
While it is inconvenient to actively notice new emulator support in
the samples (unlike the integration tests in #13799), we can still
do manual testing to remove obsolete exclusions after an SDK update.

Here we enable some of the Google SQL samples, and almost all of
the PostgreSql samples, that were previously skipped on the emulator.