ci: Bazel builds use .cloudcxxrc (#13824)
Sometimes we want to run the build scripts on our workstations, but we
are only interested in storage, or pubsub, or spanner. We can do this
for CMake-builds using `.cloudcxxrc` and the `development.sh` build.
With this change any Bazel-based build can be configured to compile a
subset of the targets.