feat(cmake): automatically set `ctype=CORD` workarounds (#13877)
With Protobuf < v23.x (aka 4.23.x) we need to enable some ugly
workarounds to use `ctype=CORD` fields in the storage library. With this
change CMake automatically sets the workaround option if needed. Bazel
assumes the version of Protobuf is >= v23 and disables the workarounds
by default. That is reasonable because Bazel users update more often (or
at least we believe that).