ci: use pkgconf-2.2.x (#13868)
In some platforms the default `pkg-config` or `pkgconf` version is
broken. They either take too long (minutes) to analyze the dependencies
of something like `Abseil`, and/or produce dependencies in the wrong
order. The 2.2.x series of `pkgconf` seems to solve these problems. It
seems better to use a newer version of `pkgconf` rather than using an
old version of `pkg-config`.