ci: modify ctest invocation in ctest::has_no_tests()
Move to `env -C "${dir}" ctest" instead of `ctest --test-dir "${dir}"`.

Not only does it hide the `Testing/Temporary/LastTest.log` file away
in the build directory, but it also seems to produce the "Total Tests"
output we're looking for, which means the cmake-oldest-deps build will
now actually run tests.