test(spanner): scope transactions within connections (#14007)
Always(*) create test transactions within the lifetime of the connection
that uses them so that those transactions will not hold references to
any sessions allocated from the test connection's pool when the pool
is destroyed.  This will make a difference after an upcoming change to
verify the final state of the test session pool.

(*) Note that there is one test, TransactionOutlivesConnection, where we
explicitly do the opposite to check that having an associated session
outside the pool after its destruction does not have any ill effects.