ci(gcb): reduce overhead of builds
To handle (what I think are rare) overload conditions we cancel a build
if a newer build for the same PR is pending. We only do this at the
beginning of the build, so it only helps if the queue is pretty deep
already. I think a better way to deal with that problem is to get more
quota.