fix(spanner): correct use of Interval offset in Timestamp addition
When adding an `Interval` to a `Timestamp`+time-zone, the offset should
be interpreted as an absolute value rather than a civil one.  (This is a
poor choice in my opinion, but so be it.)

Update the test case, which now has to use a "1 day" interval to show the
difference between `Add(ts, "1 day", tz)` and `Add(ts, "24 hours", tz)`
over a civil-time discontinuity.

Add a new test based upon the postgresql.org examples, one of which would
have demonstrated the original problem.