feat(spanner): add Interval justification
Add APIs to flow Interval 24-hour offsets into days, and 30-day periods
into months.  Seeing as "justification" is included in the relational
operators, to best way to verify the results is by conversion to a
string.

And now that we have `JustifyHours(intvl)`, apply it to the result of
`Diff(ts1, ts2, tz)` as required.