ci(gcb): move build images to artifact registry (#14142)
Practically the first step in our builds is to create (with a cache) the
Docker image with all the development tools. We have been using Google
Container Registry (GCR) to host these images and their cache. And we
have been using the "global" version of GCR.

GCR is deprecated in favor of Artifact Registry. We have been using
artifact registry as the backend of GCR. With this change we start using
artifact registry directly.

We use a regional artifact registry repository, in the same region as
the build pool. That avoids WAN traffic and hopefully gives us better
performance, lower costs, and more reliable builds.

Artifact registry has its own garbage collection, so we can simplify the
build scripts, and let artifact registry take care of removing old
images.