Another attempt at building libcxx "just so"
This worked for me on my workstation. The tests did not pass before this change, and started to pass afterwards.

I decided to use LLVM 18.1.1 because this is the version of Clang included with Fedora:40. I am a coward and fear upgrading the library beyond the corresponding version of the compiler.